Refactor provisioning process.
Components should now each have a directory under `lib/components`.
A component manifest must be provided in `viscose.yml`. An example can
be found in `lib/components/leonore/viscose.yml`. The manifest must
contain definitions for the `deploy` and `undeploy` processes as an
array of deployment steps.

Steps must have a name. A step definition can be provided inline with
the `script` attribute or be loaded from a file specified by the
`script_file` attribute. If no `script` or `script_file` attributes are
present, we expect the step definition in
`lib/components/<component-name>/<step-name>.rb`.